requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
supabase
python-dotenv